Review approaches for attributes output
Describe additional options for attribute output. The original places attribute objects in a list, then uses a separate list with attr_data_list and attr_name_list. This makes sense and works. However, the string items in attr_name_list are already in the objects in attr_data_list as the attribute_name attribute. So, it is better to refer to this attribute instead of typing a separate list. This makes the code easier to maintain because it uses less code and it is less likely that someone will introduce a bug.
The two original lists are a key and a value, a direct mapping, so they should use a map type, a dictionary. The dictionary also clarifies working with dependent values. For our email attribute, we use a given name and surname which are generated randomly. We must pass their values to the email attribute and it is more clear to call 'gname' over a list index because the order of items in the list might change.